+++ examples/sanity/app/lib/sanity/index.ts
@@ -1,5 +1,0 @@
-export * from "./config";
-export * from "./filterDataToSingleItem";
-export * from "./helpers";
-export * from "./getClient";
-export * from "./usePreviewSubscription";